pyproject.toml
[tool.poetry]
name = "amalgama"
version = "0.2.0"
description = "amalgama scraper"
authors = ["Andriy Orehov <andriyorehov@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/andriyor/amalgama-pq"

[tool.poetry.dependencies]
python = "^3.6"
pyquery = "^1.4"
python-slugify = "^2.0"

[tool.poetry.dev-dependencies]
pytest = "^4.0"
codecov = "^2.0"
pytest-cov = "^2.6"
mypy = "^0.650.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
